# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Subscriptions tracker app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of subscription tracker app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://subtracker-backend-q0d2.onrender.com"
    }
  ],
...
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "operationId": "RefreshToken",
        "description": "Generate a new access token using a refresh token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Token refreshed successfully!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/refreshTokenResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
    "/api/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user profile",
        "operationId": "getUserProfile",
        "description": "Retrieve current user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Partially update user",
        "operationId": "patchUser",
        "description": "Update specific fields of the user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/updateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete user",
        "operationId": "deleteUser",
        "description": "Delete current user profile with all subscriptions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "User account is successfully removed"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oauth.sid"
      }
    },
    "schemas": {
...
      "refreshTokenResponse": {
        "type": "object",
        "required": [
          "accessToken"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          }
        }
      },
...
      "user": {
        "type": "object",
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "68d421e4a44e5179ce84334b"
          },
          "name": {
            "type": "string",
            "example": "John"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "johndoe@email.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-02-17T10:44:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-02-17T10:44:00Z"
          },
          "avatar": {
            "type": "string",
            "example": "https://res.cloudinary.com/dfwbmillk/image/upload/v1765392797/vjzjgpat1jhkyeomwguu.gif"
          }
        }
      },
      "updateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User's name",
            "example": "Jason"
          },
          "avatar": {
            "type": "string",
            "format": "binary",
            "description": "Name and avatar file to update"
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "\"email\" must be a valid email",
                      "enum": [
                        "\"password\" is not allowed to be empty",
                        "\"email\" must be a valid email",
                        "\"email\" is not allowed to be empty",
                        "\"name\" length must be at least 3 characters long",
                        "\"name\" is not allowed to be empty",
                        "\"term\" with value \"invalid value\" fails to match the valid mongo id pattern",
                        "\"category\" with value \"invalid value\" fails to match the valid mongo id pattern",
                        "\"endDate\" must be a valid date",
                        "\"price\" must be greater than or equal to 0",
                        "File too large",
                        "Only .jpg, .jpeg, .png, .gif files are allowed"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized Error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired",
                      "enum": [
                        "User not found",
                        "Unauthorized",
                        "Please provide Authorization header",
                        "Auth header should be of type Bearer",
                        "Session not found",
                        "Session token expired",
                        "Access token expired",
                        "No access token"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}